requirements.txt
discord.py==2.0.1
aiosqlite==0.17.0
python-dotenv==0.21.0
easy-pil==0.1.9